I was a part of a team that have developed several applications using Struts, Spring and Hibernate together, and one of the problems that have faced us while using Hibernate was the rendering of the view. The problem is that when you retrieve an object 'a' of persistence class 'A' that has an instance 'b' of persistence class 'B', and this relation is lazily loaded, the value of 'b' will be 'null'. This will cause a "LazyInitializationException" while rendering the view (if you need the value of 'b' in the view of course).
The solution to this problem can be done by keeping the hibernate session alive until the view is rendered, and this is what Hibernate introduced as the Open Session In View Design Pattern. Since the Hibernate session will be opened, trying to retrieve 'b' in the view will cause Hibernate to go and fetch it from the DB. In a web application, this can be done through a filter/interceptor.

http://jnassef.blogspot.com/2007/05/open-session-in-view.html
Others links
http://wiki.apache.org/tapestry/SpringHibernate
http://jnassef.blogspot.com/2007/05/open-session-in-view.html
http://learningwebdev.blogspot.com/2009/06/open-session-in-view-iis7-spring-net.html


git-svn-id: http://svn.encuesta.me/trunk/encuestame@319 39e2479b-dd58-4041-9f67-69c9b71f0610